+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # keras 2.0.6 (unreleased)
 
+- Added `install_keras()` function which installs both TensorFlow and Keras
+
 - Added `serialize_model()` and `unserialize_model()` functions for saving 
   Keras models as 'raw' R objects.